restructured the module
