+++ java/module-screen-video/src/main/java/com/swe/Networking/AbstractNetworking.java
@@ -1,0 +1,7 @@
+package com.swe.Networking;
+
+public interface AbstractNetworking {
+    void SendData(byte[] data, String[] dest,int[] port);
+    void Subscribe(String name, MessageListener function);
+    void RemoveSubscription(String name);
+}
