+++ java/module-screen-video/src/main/java/com/swe/Networking/MessageListener.java
@@ -1,0 +1,6 @@
+package com.swe.Networking;
+
+@FunctionalInterface
+public interface MessageListener {
+    void ReceiveData(byte[] data);
+}
