+++ java/module-screen-video/src/main/java/com/swe/RPC/AbstractRPC.java
@@ -1,0 +1,8 @@
+package com.swe.RPC;
+
+import java.util.concurrent.CompletableFuture;
+
+public interface AbstractRPC {
+    void subscribe(String name, RProcedure func);
+    CompletableFuture<byte[]> Call(String name, byte[] args);
+}
