+++ java/module-screen-video/src/main/java/com/swe/RPC/RProcedure.java
@@ -1,0 +1,5 @@
+package com.swe.RPC;
+
+public interface RProcedure {
+    byte[] call(byte[] args);
+}
