+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/Capture/FrameCaptureListener.java
@@ -1,0 +1,10 @@
+package com.swe.ScreenNVideo.Capture;
+
+/**
+ * Interface for handling captured frames
+ */
+public interface FrameCaptureListener {
+    void onCaptureError(String error);
+    void onCaptureStarted();
+    void onCaptureStopped();
+}
