+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/CaptureManager.java
@@ -1,0 +1,12 @@
+package com.swe.ScreenNVideo;
+
+/**
+ * Interface to other modules
+ */
+public interface CaptureManager {
+    /**
+     * Start Screen And Video Caputering service
+     * Never returns.
+     */
+    void startCapture();
+}
