+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/Codec/Codec.java
@@ -1,0 +1,8 @@
+package com.swe.ScreenNVideo.Codec;
+
+import java.awt.image.BufferedImage;
+
+public interface Codec {
+    String Encode(int x, int y, int height, int weight);
+    BufferedImage Decode(String encoded_image);
+}
