+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/PatchGenerator/IHasher.java
@@ -1,0 +1,10 @@
+package com.swe.ScreenNVideo.PatchGenerator;
+
+public interface IHasher {
+    /**
+     * Computes a hash for an image patch/tile
+     * @param img->pixel data (int[][][]); (x, y)->location of the first corner; (w, h)->height and width
+     * @return hash as a long
+     */
+    long hash(int[][][] img, int x, int y, int w, int h);
+}
