+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/PatchGenerator/Stitchable.java
@@ -1,0 +1,6 @@
+package com.swe.ScreenNVideo.PatchGenerator;
+
+// Interface
+public interface Stitchable {
+    void applyOn(int[][][] canvas);
+}
