+++ java/module-screen-video/src/test/java/com/swe/ScreenNVideo/MediaCaptureManagerTest.java
