--- src/main/java/com.Comm_Uni_Cator/Networking/AbstractNetworking.java
@@ -1,7 +1,0 @@
-package com.Comm_Uni_Cator.Networking;
-
-public interface AbstractNetworking {
-    void SendData(byte[] data, String[] dest,int[] port);
-    void Subscribe(String name, MessageListener function);
-    void RemoveSubscription(String name);
-}
