--- src/main/java/com.Comm_Uni_Cator/RPC/RProcedure.java
@@ -1,5 +1,0 @@
-package com.Comm_Uni_Cator.RPC;
-
-public interface RProcedure {
-    byte[] call(byte[] args);
-}
