--- src/main/java/com.Comm_Uni_Cator/ScreenNVideo/CaptureManager.java
@@ -1,12 +1,0 @@
-package com.Comm_Uni_Cator.ScreenNVideo;
-
-/**
- * Interface to other modules
- */
-public interface CaptureManager {
-    /**
-     * Start Screen And Video Caputering service
-     * Never returns.
-     */
-    void startCapture();
-}
