--- src/main/java/com.Comm_Uni_Cator/ScreenNVideo/ConstantProvider.java
@@ -1,10 +1,0 @@
-package com.Comm_Uni_Cator.ScreenNVideo;
-
-public record ConstantProvider() {
-    static  int HASH_STRIDE=2;
-    static  String START_VIDEO_CAPTURE="startVideoCapture";
-    static  String STOP_VIDEO_CAPTURE="stopVideoCapture";
-    static  String START_SCREEN_CAPTURE="startScreenCapture";
-    static  String STOP_SCREEN_CAPTURE="stopScreenCapture";
-    static String MODULE_REMOTE_KEY="screenNVideo";
-}
